+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/SoraSignalingTransportType.kt
@@ -3,7 +3,7 @@ package jp.shiguredo.sora.sdk.channel
 /**
  * シグナリングの経路種別を表す列挙型.
  */
-enum class SoraSignalingTransport {
+enum class SoraSignalingTransportType {
     WEBSOCKET,
     DATA_CHANNEL,
 }